+++ Studr/Assets.xcassets/FontAwesome.dataset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "data" : [
+    {
+      "idiom" : "universal",
+      "filename" : "FontAwesome.otf"
+    }
+  ]
+}
