--- Studr/ViewController.swift
@@ -1,74 +1,0 @@
-//
-//  ViewController.swift
-//  Studr
-//
-//  Created by Joshua Herkness on 10/31/15.
-//  Copyright © 2015 JJR. All rights reserved.
-//
-
-import UIKit
-import Parse
-import ParseUI
-
-class ViewController: UIViewController, PFLogInViewControllerDelegate, PFSignUpViewControllerDelegate {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-        
-    }
-    
-    override func viewDidAppear(animated: Bool) {
-        
-        super.viewDidAppear(animated)
-        
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-    
-    // Mark: Parse Login
-    
-    func logInViewController(logInController: PFLogInViewController, shouldBeginLogInWithUsername username: String, password: String) -> Bool {
-        
-        if (!username.isEmpty || !password.isEmpty){
-            return true
-        } else {
-            return false
-        }
-    }
-    
-    func logInViewController(logInController: PFLogInViewController, didLogInUser user: PFUser) {
-        self.dismissViewControllerAnimated(true, completion: nil)
-    }
-    
-    func logInViewController(logInController: PFLogInViewController, didFailToLogInWithError error: NSError?) {
-        print("Failed to login")
-    }
-    
-    // Mark: Parse Sign Up
-    
-    func signUpViewController(signUpController: PFSignUpViewController, didSignUpUser user: PFUser) {
-        self.dismissViewControllerAnimated(true, completion: nil)
-    }
-    
-    func signUpViewControllerDidCancelSignUp(signUpController: PFSignUpViewController) {
-        print("User dismissed sign up")
-    }
-    
-    func signUpViewController(signUpController: PFSignUpViewController, didFailToSignUpWithError error: NSError?) {
-        print("Failed to sign up")
-    }
-    
-    // Mark: Actions
-
-    @IBAction func sideMenuToggleAction(sender: AnyObject) {
-        
-        let drawerViewController: DrawerViewController = self.view.window!.rootViewController as! DrawerViewController
-        drawerViewController.setPaneState(.Open, animated: true, allowUserInterruption: true, completion: nil)
-        
-    }
-}
-
